Finish economy consolidation: delete FlowSolver, collapse duplicate formulas onto primitives
Completes the ONTOLOGY.md § 6 consolidation pass:

- Delete flow/FlowSolver.ts (solveIteratively, calculateEfficiency,
  estimateOverhead): CorpPlanner has been the live authority since the
  flowAdapter swap. printSolutionSummary relocated to FlowEconomy; the one
  uniquely-pinned behavior (a far profitable source is mined when build-time
  is spare) migrated to the CorpPlanner test.
- Migrate the remaining duplicate-formula call-sites to economy/primitives:
  CarryCorp, ConstructionCorp, BodyBuilder (carryPartsFor), HarvestCorp,
  UpgradingCorp, CarryCorp (effectiveLife), FlowGraph (roundTripTicks).
- Delete the FlowTypes formula copies (calculateRoundTrip, calculateCarryParts,
  calculateHaulerCostPerTick, MINER_OVERHEAD_PER_TICK).
- Delete dead BodyBuilder helpers (calculateHaulingNeeds,
  calculateTankerCarryNeeded - no callers) and scripts/compare-efficiency.ts,
  which imported EdgeVariant functions removed in an earlier pass and no
  longer compiled.

394 unit tests pass; flow-handoff integration test green; webpack build clean.
