+++ src/flow/FlowEconomy.ts
@@ -29,7 +29,6 @@ import {
   SinkType
 } from "./FlowTypes";
 import { FlowGraph, createFlowGraph } from "./FlowGraph";
-import { printSolutionSummary } from "./FlowSolver";
 import { PRIORITY_PRESETS, PriorityManager } from "./PriorityManager";
 import { Node } from "../nodes/Node";
 import { NodeNavigator } from "../nodes/NodeNavigator";
@@ -52,13 +51,11 @@ export class FlowEconomy {
   /** Flow graph built from nodes */
   private graph: FlowGraph;
 
-
   /** Priority manager instance */
   private priorityManager: PriorityManager;
 
   /** Current solution (null if not yet solved) */
   private solution: FlowSolution | null;
-
 
   /** Current priority context */
   private context: PriorityContext | null;
@@ -439,6 +436,42 @@ export class FlowEconomy {
     }
 
     this.graph.debugPrint();
+  }
+}
+
+/**
+ * Debug: print a solution summary to the console.
+ * (Relocated from the retired FlowSolver - see docs/ONTOLOGY.md § 6.)
+ */
+export function printSolutionSummary(solution: FlowSolution): void {
+  console.log("\n=== Flow Solution ===");
+  console.log(`Computed at tick: ${solution.computedAt}`);
+  console.log(`Sustainable: ${String(solution.isSustainable)}`);
+  console.log("");
+  console.log("Energy Flow:");
+  console.log(`  Total Harvest:    ${solution.totalHarvest.toFixed(2)}/tick`);
+  console.log(`  Mining Overhead:  ${solution.miningOverhead.toFixed(2)}/tick`);
+  console.log(`  Hauling Overhead: ${solution.haulingOverhead.toFixed(2)}/tick`);
+  console.log(`  Total Overhead:   ${solution.totalOverhead.toFixed(2)}/tick`);
+  console.log(`  Net Energy:       ${solution.netEnergy.toFixed(2)}/tick`);
+  console.log(`  Efficiency:       ${solution.efficiency.toFixed(1)}%`);
+  console.log("");
+  console.log(`Miners: ${solution.miners.length}`);
+  console.log(`Haulers: ${solution.haulers.length} assignments`);
+  console.log(`Sink Allocations: ${solution.sinkAllocations.length}`);
+
+  if (solution.unmetDemand.size > 0) {
+    console.log("\nUnmet Demand:");
+    for (const [sinkId, unmet] of solution.unmetDemand) {
+      console.log(`  ${sinkId}: ${unmet.toFixed(2)}/tick`);
+    }
+  }
+
+  if (solution.warnings.length > 0) {
+    console.log("\nWarnings:");
+    for (const warning of solution.warnings) {
+      console.log(`  ⚠ ${warning}`);
+    }
   }
 }
