--- src/flow/FlowSolver.ts
@@ -1,578 +1,0 @@
-/**
- * FlowSolver - Priority-Weighted Flow Allocation
- *
- * Solves the flow allocation problem: given sources, sinks, and edges,
- * determine optimal miner/hauler assignments and energy distribution.
- *
- * This replaces Market.clear() with a global optimization approach.
- *
- * Algorithm:
- * 1. Assign miners to sources (one per source)
- * 2. Calculate total harvest capacity
- * 3. Calculate mining overhead
- * 4. For each sink (by priority), allocate energy from nearest sources
- * 5. Calculate hauling requirements for each allocation
- * 6. Verify system is sustainable (overhead < harvest)
- */
-import {
-  BODY_COSTS,
-  CREEP_LIFETIME,
-  FlowConstraints,
-  FlowEdge,
-  FlowProblem,
-  FlowSink,
-  FlowSolution,
-  FlowSource,
-  HaulerAssignment,
-  MINER_COST,
-  MINER_OVERHEAD_PER_TICK,
-  MINER_PARTS,
-  MinerAssignment,
-  SOURCE_ENERGY_PER_TICK,
-  SinkAllocation,
-  calculateCarryParts,
-  calculateHaulerCostPerTick
-} from "./FlowTypes";
-import { SPAWN_PARTS_PER_TICK } from "../corps/economics";
-
-/**
- * Fraction of a spawn's build-rate that mining + hauling may claim. The spawn
- * also builds upgraders, builders, reservers, so miners/haulers get only part of
- * its 1/3 parts-per-tick. Tunable; it sets how hard the spawn-time budget bites
- * before far sources start to fall out.
- */
-const MINING_BUDGET_FRACTION = 0.6;
-
-// =============================================================================
-// SOLVER CLASS
-// =============================================================================
-
-/**
- * FlowSolver computes optimal energy allocation across the network.
- *
- * The solver uses a greedy priority-based algorithm:
- * 1. Higher priority sinks get energy first
- * 2. Each sink draws from nearest sources
- * 3. Overhead (miners + haulers) is accounted for
- * 4. Remaining energy flows to lower priority sinks
- */
-export class FlowSolver {
-  /**
-   * Solve the flow allocation problem.
-   *
-   * @param problem - Flow problem with sources, sinks, edges, constraints
-   * @returns Flow solution with assignments and metrics
-   */
-  public solve(problem: FlowProblem): FlowSolution {
-    const { sources, sinks, edges, constraints } = problem;
-    const currentTick = typeof Game !== "undefined" ? Game.time : 0;
-
-    // Initialize solution
-    const solution: FlowSolution = {
-      miners: [],
-      haulers: [],
-      sinkAllocations: [],
-      totalHarvest: 0,
-      miningOverhead: 0,
-      haulingOverhead: 0,
-      totalOverhead: 0,
-      netEnergy: 0,
-      efficiency: 0,
-      unmetDemand: new Map(),
-      isSustainable: false,
-      warnings: [],
-      computedAt: currentTick
-    };
-
-    if (sources.length === 0) {
-      solution.warnings.push("No sources available");
-      return solution;
-    }
-
-    // Build edge lookup for fast access
-    const edgeMap = this.buildEdgeMap(edges);
-
-    // Find spawn sinks for miner assignments
-    const spawnSinks = sinks.filter(s => s.type === "spawn");
-    if (spawnSinks.length === 0) {
-      solution.warnings.push("No spawn sinks - cannot assign miners");
-      return solution;
-    }
-
-    // Step 1: Assign miners to sources
-    const minerAssignments = this.assignMiners(sources, spawnSinks, edgeMap, constraints);
-    solution.miners = minerAssignments;
-
-    // Step 2: Calculate total harvest and mining overhead
-    solution.totalHarvest = minerAssignments.reduce((sum, m) => sum + m.harvestRate, 0);
-    solution.miningOverhead = minerAssignments.reduce((sum, m) => sum + m.spawnCostPerTick, 0);
-
-    // Step 3: Track available energy per source
-    const sourceEnergy = new Map<string, number>();
-    for (const miner of minerAssignments) {
-      sourceEnergy.set(miner.sourceId, miner.harvestRate);
-    }
-
-    // Step 4: Allocate energy to sinks by priority
-    // Sort sinks by priority (should already be sorted, but ensure it)
-    const sortedSinks = [...sinks].sort((a, b) => b.priority - a.priority);
-
-    let totalHaulingOverhead = 0;
-    const haulerAssignments: HaulerAssignment[] = [];
-
-    for (const sink of sortedSinks) {
-      const allocation = this.allocateToSink(sink, sourceEnergy, edgeMap, spawnSinks, constraints);
-
-      solution.sinkAllocations.push(allocation);
-
-      // Track unmet demand
-      if (allocation.unmet > 0) {
-        solution.unmetDemand.set(sink.id, allocation.unmet);
-      }
-
-      // Create hauler assignments for this allocation
-      for (const flow of allocation.sourceFlows) {
-        if (flow.amount <= 0) continue;
-
-        const edge = edgeMap.get(this.edgeKey(flow.sourceId, sink.id));
-        if (!edge) continue;
-
-        // Find nearest spawn for these haulers
-        const nearestSpawn = this.findNearestSpawn(flow.sourceId, spawnSinks, edgeMap);
-
-        const haulerAssignment = this.createHaulerAssignment(
-          edge,
-          flow,
-          sink.id,
-          nearestSpawn?.id || spawnSinks[0].id,
-          constraints
-        );
-
-        haulerAssignments.push(haulerAssignment);
-        totalHaulingOverhead += haulerAssignment.spawnCostPerTick;
-      }
-    }
-
-    solution.haulers = haulerAssignments;
-    solution.haulingOverhead = totalHaulingOverhead;
-
-    // Step 5: Calculate final metrics
-    solution.totalOverhead = solution.miningOverhead + solution.haulingOverhead;
-    solution.netEnergy = solution.totalHarvest - solution.totalOverhead;
-    solution.efficiency = solution.totalHarvest > 0 ? (solution.netEnergy / solution.totalHarvest) * 100 : 0;
-
-    // Check sustainability
-    solution.isSustainable = solution.netEnergy >= 0;
-    if (!solution.isSustainable) {
-      solution.warnings.push(
-        `Economy not sustainable: overhead ${solution.totalOverhead.toFixed(
-          2
-        )} > harvest ${solution.totalHarvest.toFixed(2)}`
-      );
-    }
-
-    // Check minimum controller upgrade
-    const controllerAlloc = solution.sinkAllocations.find(a => a.sinkType === "controller");
-    if (controllerAlloc && controllerAlloc.allocated < constraints.minControllerUpgrade) {
-      solution.warnings.push(
-        `Controller upgrade ${controllerAlloc.allocated.toFixed(2)} below minimum ${constraints.minControllerUpgrade}`
-      );
-    }
-
-    return solution;
-  }
-
-  // ===========================================================================
-  // MINER ASSIGNMENT
-  // ===========================================================================
-
-  /**
-   * Assign miners to sources.
-   * Each source gets one miner, assigned to nearest spawn.
-   * Only assigns to sources that are profitable (net positive energy after overhead).
-   */
-  private assignMiners(
-    sources: FlowSource[],
-    spawnSinks: FlowSink[],
-    edgeMap: Map<string, FlowEdge>,
-    _constraints: FlowConstraints
-  ): MinerAssignment[] {
-    // Phase 1 - cost every source from its nearest spawn. A source is a CANDIDATE
-    // whenever it nets positive ENERGY (its real profit), no matter how far: we
-    // never skip a profitable source just for distance. We also record the spawn
-    // BUILD-TIME its miner+hauler fleet consumes (parts/tick over the TTL: a static
-    // miner walks out and then dies at the source, so it amortizes over
-    // CREEP_LIFETIME minus the walk). That build-time is what Phase 2 budgets.
-    interface Candidate {
-      spawnId: string;
-      netEnergy: number;
-      spawnPartsPerTick: number;
-      assignment: MinerAssignment;
-    }
-    const candidates: Candidate[] = [];
-
-    for (const source of sources) {
-      const nearestSpawn = this.findNearestSpawn(source.id, spawnSinks, edgeMap);
-      if (!nearestSpawn) continue;
-
-      const spawnDistance = nearestSpawn.distance;
-      const harvestRate = source.capacity;
-      const life = Math.max(1, CREEP_LIFETIME - spawnDistance);
-      const minerOverhead = MINER_COST / life;
-      const carryParts = calculateCarryParts(harvestRate, spawnDistance);
-      const haulerOverhead = (carryParts * (BODY_COSTS.CARRY + BODY_COSTS.MOVE)) / life;
-      const netEnergy = harvestRate - minerOverhead - haulerOverhead;
-
-      if (netEnergy <= 0) {
-        console.log(
-          `[FlowSolver] Skipping source ${source.id.slice(-8)} (net-energy-negative): ` +
-            `net=${netEnergy.toFixed(2)}, distance=${spawnDistance}`
-        );
-        continue;
-      }
-
-      const spawnPartsPerTick = (MINER_PARTS + 2 * carryParts) / life; // miner 8 + hauler CARRY+MOVE
-      candidates.push({
-        spawnId: nearestSpawn.id,
-        netEnergy,
-        spawnPartsPerTick,
-        assignment: {
-          sourceId: source.id,
-          nodeId: source.nodeId,
-          spawnId: nearestSpawn.id,
-          spawnDistance,
-          harvestRate,
-          spawnCostPerTick: minerOverhead,
-          maxMiners: source.maxMiners,
-          efficiency: (netEnergy / harvestRate) * 100
-        }
-      });
-    }
-
-    // Phase 2 - the DYNAMIC spawn-time wall. Each spawn builds one part every 3
-    // ticks; mining+hauling may claim only a fraction of that (consumers need the
-    // rest). Per spawn, keep candidates by descending net energy PER build-part,
-    // spending the budget; the rest fall out. So a far source is dropped ONLY when
-    // its spawn is actually the bottleneck - when there is spare build-time, every
-    // profitable source is mined no matter how far. This is the spawn-time wall as
-    // CONTENTION, not a fixed price: an idle spawn reaches far, a saturated one
-    // pulls in, and the cutoff distance emerges from the sources actually present.
-    const budgetPerSpawn = SPAWN_PARTS_PER_TICK * MINING_BUDGET_FRACTION;
-    candidates.sort((a, b) => b.netEnergy / b.spawnPartsPerTick - a.netEnergy / a.spawnPartsPerTick);
-
-    const usedBySpawn = new Map<string, number>();
-    const assignments: MinerAssignment[] = [];
-    for (const c of candidates) {
-      const spent = usedBySpawn.get(c.spawnId) ?? 0;
-      // Always staff a spawn's first (best) source even if it alone exceeds the
-      // mining budget - a spawn with one far source should still mine it.
-      if (spent > 0 && spent + c.spawnPartsPerTick > budgetPerSpawn) {
-        console.log(
-          `[FlowSolver] Source ${c.assignment.sourceId.slice(-8)} falls out: ` +
-            `spawn build-time budget spent (distance=${c.assignment.spawnDistance})`
-        );
-        continue;
-      }
-      usedBySpawn.set(c.spawnId, spent + c.spawnPartsPerTick);
-      assignments.push(c.assignment);
-    }
-
-    return assignments;
-  }
-
-  // ===========================================================================
-  // SINK ALLOCATION
-  // ===========================================================================
-
-  /**
-   * Allocate energy to a single sink from available sources.
-   * Uses greedy nearest-source allocation.
-   */
-  private allocateToSink(
-    sink: FlowSink,
-    sourceEnergy: Map<string, number>,
-    edgeMap: Map<string, FlowEdge>,
-    _spawnSinks: FlowSink[],
-    _constraints: FlowConstraints
-  ): SinkAllocation {
-    const allocation: SinkAllocation = {
-      sinkId: sink.id,
-      sinkType: sink.type,
-      allocated: 0,
-      demand: sink.demand,
-      unmet: sink.demand,
-      priority: sink.priority,
-      sourceFlows: []
-    };
-
-    // Special case: spawn sink needs overhead allocation
-    // This is the energy to produce miners/haulers, not direct transfer
-    if (sink.type === "spawn") {
-      // Spawn overhead is accounted for in mining overhead
-      // Here we just need to ensure energy reaches the spawn for creep production
-      // The "demand" for spawn is the overhead we need to cover
-    }
-
-    // Get all edges to this sink, sorted by distance
-    const edgesToSink: { sourceId: string; edge: FlowEdge; available: number }[] = [];
-
-    for (const [sourceId, available] of sourceEnergy) {
-      if (available <= 0) continue;
-
-      const edge = edgeMap.get(this.edgeKey(sourceId, sink.id));
-      if (edge) {
-        edgesToSink.push({ sourceId, edge, available });
-      }
-    }
-
-    // Sort by distance (greedy nearest-first)
-    edgesToSink.sort((a, b) => a.edge.distance - b.edge.distance);
-
-    // Allocate from nearest sources until demand is met or sources exhausted
-    let remaining = Math.min(sink.demand, sink.capacity);
-
-    for (const { sourceId, edge, available } of edgesToSink) {
-      if (remaining <= 0) break;
-
-      const take = Math.min(available, remaining);
-
-      // Deduct from source
-      sourceEnergy.set(sourceId, available - take);
-
-      // Record flow
-      allocation.sourceFlows.push({
-        sourceId,
-        amount: take,
-        distance: edge.distance
-      });
-
-      allocation.allocated += take;
-      remaining -= take;
-    }
-
-    allocation.unmet = Math.max(0, sink.demand - allocation.allocated);
-
-    return allocation;
-  }
-
-  // ===========================================================================
-  // HELPER METHODS
-  // ===========================================================================
-
-  /**
-   * Build edge lookup map for fast access.
-   */
-  private buildEdgeMap(edges: FlowEdge[]): Map<string, FlowEdge> {
-    const map = new Map<string, FlowEdge>();
-    for (const edge of edges) {
-      map.set(edge.id, edge);
-      // Also index by from|to key
-      map.set(this.edgeKey(edge.fromId, edge.toId), edge);
-    }
-    return map;
-  }
-
-  /**
-   * Create edge key for lookup.
-   */
-  private edgeKey(fromId: string, toId: string): string {
-    return fromId < toId ? `${fromId}|${toId}` : `${toId}|${fromId}`;
-  }
-
-  /**
-   * Find nearest spawn to a source.
-   */
-  private findNearestSpawn(
-    sourceId: string,
-    spawnSinks: FlowSink[],
-    edgeMap: Map<string, FlowEdge>
-  ): { id: string; distance: number } | null {
-    let nearest: { id: string; distance: number } | null = null;
-
-    for (const spawn of spawnSinks) {
-      const edge = edgeMap.get(this.edgeKey(sourceId, spawn.id));
-      if (edge && (!nearest || edge.distance < nearest.distance)) {
-        nearest = { id: spawn.id, distance: edge.distance };
-      }
-    }
-
-    return nearest;
-  }
-
-  /** Create a hauler assignment with the classic 1:1 CARRY:MOVE sizing. */
-  private createHaulerAssignment(
-    edge: FlowEdge,
-    flow: { sourceId: string; amount: number; distance: number },
-    sinkId: string,
-    spawnId: string,
-    _constraints: FlowConstraints
-  ): HaulerAssignment {
-    const carryParts = calculateCarryParts(flow.amount, flow.distance);
-    const haulerCost = calculateHaulerCostPerTick(carryParts);
-
-    return {
-      edgeId: edge.id,
-      fromId: flow.sourceId,
-      toId: sinkId,
-      distance: flow.distance,
-      carryParts,
-      flowRate: flow.amount,
-      spawnCostPerTick: haulerCost,
-      spawnId
-    };
-  }
-}
-
-// =============================================================================
-// ITERATIVE SOLVER (For Bootstrap/Deficit Scenarios)
-// =============================================================================
-
-/**
- * Iteratively solve flow allocation, adjusting for overhead feedback.
- *
- * The basic solver doesn't account for the circular dependency:
- * hauling overhead depends on allocations, but allocations depend on
- * available energy after overhead.
- *
- * This iterative solver:
- * 1. Makes initial allocation ignoring hauling overhead
- * 2. Calculates actual hauling overhead
- * 3. Adjusts allocations to account for overhead
- * 4. Repeats until convergence
- *
- * @param problem - Flow problem
- * @param maxIterations - Maximum iterations (default: 10)
- * @returns Converged flow solution
- */
-export function solveIteratively(problem: FlowProblem, maxIterations = 10): FlowSolution {
-  const solver = new FlowSolver();
-  let solution = solver.solve(problem);
-
-  // If already sustainable, return
-  if (solution.isSustainable && solution.unmetDemand.size === 0) {
-    return solution;
-  }
-
-  // Iteratively reduce allocations to account for overhead
-  const originalDemands = problem.sinks.map(s => s.demand);
-  for (let i = 0; i < maxIterations; i++) {
-    const prevOverhead = solution.totalOverhead;
-
-    // Fit sink demands to the energy left after mining overhead. Do it by
-    // PRIORITY, not proportionally: fill the highest-priority sinks to their full
-    // demand and starve the lowest-priority ones, so scarcity cuts the least
-    // valuable work first (e.g. upgrading yields to building during construction).
-    // Scaling every demand by the same factor - the old behavior - quietly
-    // inverted the priorities, letting a big low-priority demand (the controller)
-    // out-draw a smaller high-priority one (construction).
-    const availableForSinks = solution.totalHarvest - solution.miningOverhead;
-    const totalDemand = originalDemands.reduce((sum, d) => sum + d, 0);
-
-    // Restore original demands each pass so a transient overhead spike can't
-    // ratchet them permanently downward.
-    problem.sinks.forEach((s, idx) => (s.demand = originalDemands[idx]));
-
-    if (availableForSinks < totalDemand) {
-      let budget = Math.max(0, availableForSinks);
-      for (const sink of [...problem.sinks].sort((a, b) => b.priority - a.priority)) {
-        if (sink.demand <= budget) {
-          budget -= sink.demand;
-        } else {
-          sink.demand = budget;
-          budget = 0;
-        }
-      }
-    }
-
-    // Re-solve
-    solution = solver.solve(problem);
-
-    // Check convergence
-    if (Math.abs(solution.totalOverhead - prevOverhead) < 0.01) {
-      break;
-    }
-  }
-
-  return solution;
-}
-
-// =============================================================================
-// UTILITY FUNCTIONS
-// =============================================================================
-
-/**
- * Calculate efficiency for a given harvest and overhead.
- */
-export function calculateEfficiency(harvest: number, overhead: number): number {
-  if (harvest <= 0) return 0;
-  return ((harvest - overhead) / harvest) * 100;
-}
-
-/**
- * Estimate total overhead for a flow problem.
- * Quick estimate without full solve.
- */
-export function estimateOverhead(problem: FlowProblem): {
-  miningOverhead: number;
-  haulingOverhead: number;
-  totalOverhead: number;
-} {
-  const miningOverhead = problem.sources.length * MINER_OVERHEAD_PER_TICK;
-
-  // Estimate hauling based on average distance
-  let totalHaulDistance = 0;
-  let edgeCount = 0;
-
-  for (const edge of problem.edges) {
-    totalHaulDistance += edge.distance;
-    edgeCount++;
-  }
-
-  const avgDistance = edgeCount > 0 ? totalHaulDistance / edgeCount : 30;
-  const totalHarvest = problem.sources.length * SOURCE_ENERGY_PER_TICK;
-
-  // Rough estimate: all energy hauled at average distance
-  const avgCarryParts = calculateCarryParts(totalHarvest, avgDistance);
-  const haulingOverhead = calculateHaulerCostPerTick(avgCarryParts);
-
-  return {
-    miningOverhead,
-    haulingOverhead,
-    totalOverhead: miningOverhead + haulingOverhead
-  };
-}
-
-/**
- * Debug: Print solution summary.
- */
-export function printSolutionSummary(solution: FlowSolution): void {
-  console.log("\n=== Flow Solution ===");
-  console.log(`Computed at tick: ${solution.computedAt}`);
-  console.log(`Sustainable: ${String(solution.isSustainable)}`);
-  console.log("");
-  console.log("Energy Flow:");
-  console.log(`  Total Harvest:    ${solution.totalHarvest.toFixed(2)}/tick`);
-  console.log(`  Mining Overhead:  ${solution.miningOverhead.toFixed(2)}/tick`);
-  console.log(`  Hauling Overhead: ${solution.haulingOverhead.toFixed(2)}/tick`);
-  console.log(`  Total Overhead:   ${solution.totalOverhead.toFixed(2)}/tick`);
-  console.log(`  Net Energy:       ${solution.netEnergy.toFixed(2)}/tick`);
-  console.log(`  Efficiency:       ${solution.efficiency.toFixed(1)}%`);
-  console.log("");
-  console.log(`Miners: ${solution.miners.length}`);
-  console.log(`Haulers: ${solution.haulers.length} assignments`);
-  console.log(`Sink Allocations: ${solution.sinkAllocations.length}`);
-
-  if (solution.unmetDemand.size > 0) {
-    console.log("\nUnmet Demand:");
-    for (const [sinkId, unmet] of solution.unmetDemand) {
-      console.log(`  ${sinkId}: ${unmet.toFixed(2)}/tick`);
-    }
-  }
-
-  if (solution.warnings.length > 0) {
-    console.log("\nWarnings:");
-    for (const warning of solution.warnings) {
-      console.log(`  ⚠ ${warning}`);
-    }
-  }
-}
